fsc watchdog removal.
In addition, fixed a potential initialization problem
in the fsc server, which would open a socket, get a port
number, and close the socket, only to try to reallocate
the same port number a bit later, even though the port
might have become unavailable in the meantime. Now the
socket is initialized only once, and the port is not
released.



git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@16313 5e8d7ff9-d8ef-0310-90f0-a4852d11357a