While working on Any.## I ran across some interesting tests being made in TreeBuilder:
  val buf = new ListBuffer[(Name, Tree, Position)]
  [...]
  if (buf.iterator forall (name !=)) ...

This is always true because a Name will never equal a Tuple3.
Oh universal equality, will you never tire of toying with us?
Given that this bug has existed since r12886 one might reasonably
question the necessity of the conditionals falling prey to this,
but leaving that for another day, it should at least check what
it's trying to check.  No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@21010 5e8d7ff9-d8ef-0310-90f0-a4852d11357a