Couldn't bear to see sbt seemingly so close but still unable to run tests.  Made some progress but bogged down in scalacheck when I realized things were farther from working than I'd thought. Please forgive this brief expression of frustration, because the despair is palpable: the weeks of work I did on partest was apparently of no interest, and now I'm in here having to fix the same partest bugs again, in a more virulent form.  They've evolved into some kind of ant-partest-sbt artificial creature, more fearsome than any one of them could be under its own power.
Did we not know we have a whole directory full of pretty code like this?

  List(
    "Scala compiler classes in: " + testBuildDir,
    "Scala version is:          " + nsc.Properties.versionMsg,
    "Scalac options are:        " + universe.scalacOpts,
    "Java binaries in:          " + javaBin,
    "Java runtime is:           " + javaInfoString,
    "Java runtime options:      " + (Process.javaVmArguments mkString " "),
    "Javac options are:         " + universe.javacOpts,
    "Java options are:          " + universe.javaOpts,
    "Source directory is:       " + src,
    "Selected categories:       " + (selectedCategories mkString " "),
    ""
  ) mkString "\n"

It's src/partest-alternative in case anyone wants to consider any of it.
At the very least there are individual files or functions which could be
cherry picked.

-- This patch contains: --

Readability efforts.  Twelve parameter functions whose last six
parameters are "true,true,failedOnly,true,false,log" are pretty
challenging to work with.

Changes to the environment variables sbt was setting to look a bit
more like what partest expects and will crash without.

Sets partest.debug=true if the sbt logger level is at Debug.

On my machine all test groups except scalacheck appear to run
to completion.

Review by moix.  (Hello, moix! I am available for assistance if you
are in the market for it.  You have brought us long suffering ant-ers
to the brink of sbt goodness: and now, we enter the fray!)

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@22976 5e8d7ff9-d8ef-0310-90f0-a4852d11357a