Restoring negative literal parsing behavior to what should be the least surprising option.  Thanks much to Johannes Rudolph for identifying the bug in the bytecode generator which needed addressing for us to arrive at proper -0.0 behavior, and for writing the majority of this patch.
A '-' followed immediately by either a number or a period should
now always be treated as a single numeric literal, which means the
minus binds more tightly than anything else.  A specific example
of how this differs from 2.8 final is:

  -5.+(10) == 5.0   // and not -15.0

The full range of potentially ambiguous parses involving prefix
operators, numbers, and dots is quite large and still needs to be
completely and clearly specified.

Closes #2378 and #3657, review by odersky, jrudolph.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23024 5e8d7ff9-d8ef-0310-90f0-a4852d11357a