closes #3281, #3866. !!! rollback constraints accumulated during isCoercible check. also had to disable clearing of the undoLog.
I don't understand why we had to clear() the undoLog in the first place, since the undoXXX methods increase and decrease its size symmetrically, so the log should always be empty once they have all unwound. Was it a (premature) optimisation or was there some kind of semantic meaning to it that I didn't see?

review by odersky

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23244 5e8d7ff9-d8ef-0310-90f0-a4852d11357a