Fix 'Symbol.companionModule' for the resident mode compiler. It was confused by having modules being translated to lazy values. The direct consequence was a crash in the build manager when looking at a constructor using a default argument (compiled separately), but only on the second run.
The resident compiler may run many times over, and symbols may be reused. Therefore, a
module symbol that has been translated to a lazy val by refchecks is not guaranteed to 
have MODULE set on the next run (even before refcheck). Flags are not part of symbol
history. Instead we rely on the fact that a synthetic lazy value must have been a 
module.

review by odersky.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23825 5e8d7ff9-d8ef-0310-90f0-a4852d11357a