Continuing to press on RefChecks, discovered what seemed a pretty spectacular performance bug, at least at first.  I was disappointed to see that the time gain didn't measure up. Still, nanos are nanos.
According to YourKit, millis spent in List#foreach on a compilation
of src/library broke down like this:

|  +---LinearSeqOptimized$class.foreach               |  173,123  100 %  |
|    +---RefChecks$RefCheckTransformer.register$1     |  117,658   68 %  |

I think that is largely a profiler lie, but there is enough truth.
After adjusting to prune before recursing, the number of invocations
of method register in validateBaseTypes drops by 98%.

Review by odersky.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25164 5e8d7ff9-d8ef-0310-90f0-a4852d11357a