A conceivably pretty bad performance bug in builders.
SI-4821 pointed out that ArrayBuffer's ++ checks for a cheap size
method by matching on IndexedSeq, but mutable.IndexedSeq, so all
immutable collections are thrown in the same group as linear seqs.
I went looking for other examples of this and found them, in
key classes like Builder.

The "type shadowing trap" is a serious issue in the collections.
Closes SI-4821, no review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25509 5e8d7ff9-d8ef-0310-90f0-a4852d11357a