Add ability to document the root package
When the index of a generated set of scaladoc documentation is loaded,
the root package is displayed by default. This page would be a useful
space in which to provide some introductory documentation and links to
important packages / types. This content should however be customisable
for documentation generated for different libraries, therefore a new
property has been added to the scaladoc tool (--doc-root-content) to
specify the file from which the content should come from.

Basic root documentation, to provide a starting point, has been added
for the standard library.