Fixes #4351.
Added an "Abstract" method info to the specialized phase, which denotes that no implementation should be generated.

Previously:

trait A[@specialized(Boolean) T] {
  def foo: T
}

used to generate:

trait A$mcZ$sp extends A[Boolean] {
  def foo$mcZ$sp = this.foo
}

which caused cyclic calls because of linearization rules when several traits are mixed together. Now, the following is generated:

trait A$mcZ$sp extends A[Boolean] {
  def foo$mcZ$sp: Boolean
}

Review by dragos.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25722 5e8d7ff9-d8ef-0310-90f0-a4852d11357a