Forces classes and traits extending scala.Dynamic to define applyDynamic.
Fixes SI-4536 - Dynamic is broken.

This second fix includes two changes:

(1) Preventing the compiler running away by stopping it dynamically applying
a call to "applyDynamic". This is sufficient to stop this bug proper, but puts
the weight on the client, as the error only shows up at the point of call.

(2) Forbidding any subclassing of scala.Dynamic without declaring an
applyDynamic member – even traits or abstract classes.

Notes:

* Change (2) is necessary because forbidding it only for concrete classes is not
sufficient because intermediate abstract classes would cause the error in (1)
to come up, which could be pretty mysterious if it's a library class and the
user doesn't even know about Dynamic.

* Change (2) doesn't limit the usefulness of Dynamic, as there is no reason to
mix that trait into the hierarchy until applyDynamic is defined.

* With change (2) alone, the only way to run into SD-4536 would be to call a
dynamic method on scala.Dynamic directly or on class file compiled with an older
version of the compiler. It's still worth having (1) as a guard, though.

* This change includes tests, and the whole suite has been verified to pass,
but it is my first proposal for a change in the compiler, so please review it
carefully.