Reflection classloader adjustment.
With this change, one can use the reflective compiler with the
scala libs on the boot classpath (the default) or not (with -nobootcp)
and it will work either way.  Maybe it will work in sbt now too, but
I haven't tried it.  Since I bet you will try it, review by grek.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25932 5e8d7ff9-d8ef-0310-90f0-a4852d11357a