More repl work.
Hardened the repl against a bunch of erroneous conditions.
Added a :reset command which causes the repl to forget everything
you have told it.  Added classloader tracing when repl tracing is
enabled.  :reset is not that useful in its current form but
it's the precursor to something more interesting.  No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25941 5e8d7ff9-d8ef-0310-90f0-a4852d11357a