Fixing checkinit build.
As near as I can tell, -Xcheckinit was failing because a var was read
too early as null rather than waiting for it to be set to its initial
value, null. I guess that's a defensible warning, since one might change
it away from null. Organized the vars in that file better.  No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25954 5e8d7ff9-d8ef-0310-90f0-a4852d11357a