Allow to load $class classes using Scala reflection.
Tweaked implementation of invalidClassName method
to exclude *$class clasess from the set of invalid
names. It's not exactly clear what was the intent
of this method in first place so I'm not sure if
it's the best way to fix SI-5176. Added test-case
that covers this issue.

Fixes SI-5176. Review by odersky.



git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25982 5e8d7ff9-d8ef-0310-90f0-a4852d11357a