Fix for what have been rather uncommon common owners.
The complete histogram of results for "commonOwner" until this
patch, when building quick.lib and quick.comp:

  Calculated common owner       Occurrences
  -----------------------       -----------
  NoSymbol                      299,242
  Everything Else               0

Since I'm always paranoid somebody will think I'm the one who
broke such things in the first place, I got in the gitmobile and
fingered the responsible party.

Looks OK when it was checked in: r3930, Feb 4 2005. And it was smooth
sailing until... r4026, Mar 22 2005. So 6 1/2 weeks of goodness for poor
commonOwnerMap, to be followed by 6 1/2 years of endless NoSymboldom. In
2005 I was still making a living grifting strangers in seedy gambling
halls, so I think I'm in the clear. Here's the exact spot.

  https://github.com/scala/scala/commit/ae0da87d1aaeded9eb7f9c9ecbed8a313667757d#L12L991

I found this while trying to figure out why we are generating so many
refinements. This doesn't fix any of that, but maybe takes us a notch
closer. Review by odersky.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@26038 5e8d7ff9-d8ef-0310-90f0-a4852d11357a