Implemented manifest-based class-paths.
If you run a jar directly, like

  scala foo.jar

Then if a Class-Path attribute is present in the jar manifest, the
classpath will be constructed from that instead of the arguments. Some
things remain to be determined, like whether it's supposed to replace
a classpath given on the command line or supplement it, and whether
the master jar should be on the classpath or only and exactly the jars
listed in the manifest.

There's a really nice test case, which won't be run of course, but I
can't stand going any further without tests for these hard to test on
all platforms things.  The faux .check file shows what I see.

Closes SI-4355, review by harrah.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@26041 5e8d7ff9-d8ef-0310-90f0-a4852d11357a