Fixes for 3 tests which aren't failing.
Which should be failing, and would be failing if they were being
compiled by trunk, since they used methods which don't exist. Why aren't
they failing? That is the real mystery. In the interests of removing a
disincentive to track that one down, I future-proofed them against that
day when they start being run like they're supposed to be.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@26090 5e8d7ff9-d8ef-0310-90f0-a4852d11357a